fix CORS for flutter web
With this change appOrigin doesn't do an exact match, instead it returns the client's origin header if it starts with any of the appOrigins. i.s. A source port in the url string does not interfere with matching. The Origin header cannot be set in a chrome-js XmlHttpRequest. It always sends "http://localhost:<port>" and the source port varies.  Other targets can set Origin header directly.  This if you think web could be useful for mobile app development.